Move countries to separate package
