+++ go/data-source-sql/server.go
@@ -6,6 +6,7 @@ import (
 	"encoding/json"
 	"net/http"
 	"strings"
+	"github.com/signavio/connector-examples/go/data-source-sql/countries"
 )
 
 var address = ":9000"
